BREAKING CHANGE: update onResultClickedFn
