+++ packages/search-ui/src/Results/types.ts
@@ -1,4 +1,5 @@
 import { BoxProps, ImageProps } from '@sajari/react-components';
+import { ResultValues } from '@sajari/react-hooks';
 
 export type ColumnValue = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
 export type GapValue = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
@@ -15,14 +16,4 @@ export interface ResultsProps extends BoxProps {
   gap?: GapValue | Record<number, GapValue>;
 }
 
-export interface ResultValues {
-  _id: string;
-  url: string;
-  title: string;
-  subtitle?: string;
-  description?: string;
-  image?: string | Array<string>;
-  rating?: number;
-  price?: string | Array<string>;
-  originalPrice?: string | Array<string>;
-}
+export { ResultValues };
